refactor folder helpers > card to helpers > components
