+++ src/cards/climates/card.js
@@ -1,8 +1,8 @@
 import {LitElement, html, nothing} from 'lit';
 import {isEqual} from 'lodash-es';
 
-import '../../helpers/card/tiles.js';
-import '../../helpers/card/toast.js';
+import '../../helpers/components/tiles.js';
+import '../../helpers/components/toast.js';
 import common_style from '../../helpers/common_style.js';
 import '../../helpers/entities/climate.js';
 import {ENTITY_KIND_CLIMATE} from '../../helpers/entities/climate_const.js';
