+++ src/cards/lights/editor.js
@@ -1,6 +1,6 @@
 import {html, nothing} from 'lit';
 
-import {SciFiBaseEditor} from '../../helpers/card/base_editor.js';
+import {SciFiBaseEditor} from '../../helpers/components/base_editor.js';
 import common_style from '../../helpers/common_style.js';
 import editor_common_style from '../../helpers/editor_common_style.js';
 import {House} from '../../helpers/entities/house.js';
