+++ src/cards/weather/card.js
@@ -2,7 +2,7 @@ import Chart from 'chart.js/auto';
 import {LitElement, html, nothing} from 'lit';
 import {isEqual} from 'lodash-es';
 
-import '../../helpers/card/tiles.js';
+import '../../helpers/components/tiles.js';
 import common_style from '../../helpers/common_style.js';
 import {
   DailyForecast,
