+++ src/helpers/components/base_editor.js
