+++ src/helpers/components/stack_bar.js
